refactor: use React.lazy for components/Dialog [YTFRONT-3814]
