+++ packages/ui/src/ui/components/Dialog/Dialog.scss
@@ -1,10 +1,4 @@
 .yt-dialog {
-    &__error {
-        margin: 0 -32px;
-        padding: 20px 32px;
-        overflow: auto;
-    }
-
     &__as-block {
         display: inline-block;
         margin-left: -32px;
