+++ packages/ui/src/ui/components/Dialog/DialogError.scss
@@ -1,0 +1,5 @@
+.yt-dialog-error {
+    margin: 0 -32px;
+    padding: 20px 32px;
+    overflow: auto;
+}
