+++ packages/ui/src/ui/components/Dialog/controls/RoleListControl/RoleListControl.tsx
@@ -1,9 +1,4 @@
 import React from 'react';
-import cn from 'bem-cn-lite';
-
-import filter_ from 'lodash/filter';
-import forEach_ from 'lodash/forEach';
-import map_ from 'lodash/map';
 
 import {
     EditableListItemType,
@@ -14,12 +9,12 @@ import {
 import {SubjectCard} from '../../../SubjectLink/SubjectLink';
 import {ResponsibleType, RoleConverted} from '../../../../utils/acl/acl-types';
 import SubjectsControl from '../../../../containers/ACL/SubjectsControl/SubjectsControl';
+
 import './RoleListControl.scss';
-import {PreparedRole} from '../../../../utils/acl';
 
-const block = cn('role-list-control');
+import {block} from './utils';
 
-interface Props {
+export interface RoleListControlProps {
     className?: string;
     value: {
         newItems: Array<ResponsibleType>;
@@ -29,13 +24,13 @@ interface Props {
         unrecognized: EditableManyListsItemType<RoleConverted>;
     };
 
-    onChange: (value: Props['value']) => void;
+    onChange: (value: RoleListControlProps['value']) => void;
     placeholder?: string;
     maxVisibleCount: number;
     allowedTypes?: Array<'users' | 'groups'>;
 }
 
-export default class RoleListControl extends React.Component<Props> {
+export default class RoleListControl extends React.Component<RoleListControlProps> {
     static defaultProps = {
         maxVisibleCount: 3,
     };
@@ -54,14 +49,14 @@ export default class RoleListControl extends React.Component<Props> {
         return false;
     }
 
-    static prepareManyListData(value: Props['value']) {
+    static prepareManyListData(value: RoleListControlProps['value']) {
         const {current, toAdd, toRemove, unrecognized} = value;
         return [{...current}, toAdd, toRemove, unrecognized].filter(
             (item) => item.data && item.data.length > 0,
         );
     }
 
-    onSubejctsControlChange = (newItems?: Props['value']['newItems']) => {
+    onSubejctsControlChange = (newItems?: RoleListControlProps['value']['newItems']) => {
         const {value, onChange} = this.props;
         onChange({
             ...value,
@@ -114,98 +109,4 @@ export default class RoleListControl extends React.Component<Props> {
             </div>
         );
     }
-}
-
-export function prepareRoleListValue(roles: Array<PreparedRole>, otherMembers: Array<string> = []) {
-    const current: typeof roles = [];
-    const toAdd: typeof roles = [];
-    const toRemove: typeof roles = [];
-    const unrecognized: typeof roles = [];
-    forEach_(roles, (item) => {
-        const {isUnrecognized, isRequested, isApproved, isDepriving} = item;
-        if (isUnrecognized) {
-            unrecognized.push(item);
-        } else if (isDepriving) {
-            toRemove.push(item);
-        } else if (isApproved || isRequested) {
-            toAdd.push(item);
-        } else {
-            current.push(item);
-        }
-    });
-
-    const currentArr = current.map((item) => prepareItemOfCurrent(item));
-    const otherArr = otherMembers.map((item) => ({
-        title: item,
-        data: {},
-        frozen: true,
-    }));
-
-    return {
-        newItems: [],
-        current: {title: 'Current', data: currentArr.concat(otherArr)},
-        toAdd: {
-            title: 'Will be added',
-            itemClassName: block('item-to-add'),
-            data: toAdd.map((item) => prepareItemOfCurrent(item, {frozen: true})),
-        },
-        toRemove: {
-            title: 'Will be removed',
-            itemClassName: block('item-to-remove'),
-            data: toRemove.map((item) => prepareItemOfCurrent(item, {frozen: true})),
-        },
-        unrecognized: {
-            title: 'Unrecognized',
-            itemClassName: block('item-unrecognized'),
-            data: unrecognized.map((item) => prepareItemOfCurrent(item, {frozen: true})),
-        },
-    };
-}
-
-function prepareItemOfCurrent(role: PreparedRole, extraProps: any = {}) {
-    return {
-        title: role.text || role.value,
-        data: role,
-        ...extraProps,
-    };
-}
-
-export function roleListValueToSubjectList(value: Props['value']): Array<ResponsibleType> {
-    const {current, newItems, toAdd} = value;
-    return [
-        ...newItems,
-        ...manyListDataItemToSubjectList(current),
-        ...manyListDataItemToSubjectList(toAdd),
-    ];
-}
-
-function manyListDataItemToSubjectList(
-    manyListDataItem: EditableManyListsItemType<RoleConverted>,
-): Array<ResponsibleType> {
-    const {data} = manyListDataItem || {};
-    return map_(
-        filter_(data, ({removed}) => !removed),
-        ({data}) => {
-            const {type, value} = data || {};
-            return {type: type!, value: value!};
-        },
-    );
-}
-
-export function extractChangedSubjects(value: {
-    current: EditableManyListsItemType<RoleConverted>;
-    newItems: Array<ResponsibleType>;
-}) {
-    const {current, newItems} = value;
-    const added = newItems || [];
-    const removed = ((current && current.data) || []).filter(({removed}) => removed);
-    return {
-        added: added.map(({type, value}) => {
-            return type === 'users' ? {user: value} : {group: value};
-        }),
-        removed: removed.map(({data}) => {
-            const {value, type} = data || {};
-            return type === 'users' ? {user: value} : {group: value};
-        }),
-    };
 }
