+++ packages/ui/src/ui/pages/navigation/content/Table/TableMeta/AutomaticModeSwitch.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {Switch} from '@gravity-ui/uikit';
 
 import StatusBulb from '../../../../../components/StatusBulb/StatusBulb';
-import {FormApi, YTDFDialog, makeErrorFields} from '../../../../../components/Dialog/Dialog';
+import {FormApi, YTDFDialog, makeErrorFields} from '../../../../../components/Dialog';
 import {YTError} from '../../../../../types';
 
 export interface Props {
