+++ packages/ui/src/ui/pages/navigation/modals/TableMergeSortModal/TableSortModal.tsx
@@ -3,7 +3,7 @@ import cn from 'bem-cn-lite';
 
 import map_ from 'lodash/map';
 
-import {DialogError, DialogField, YTDFDialog} from '../../../../components/Dialog/Dialog';
+import {DialogError, DialogField, YTDFDialog} from '../../../../components/Dialog';
 import {useDispatch, useSelector} from 'react-redux';
 import {
     getNavigationTableAttributesValues,
