+++ packages/ui/src/ui/pages/users/UsersPageEditor/UsersPageEditor.tsx
@@ -1,4 +1,4 @@
-import {FormApi, YTDFDialog, makeErrorFields} from '../../../components/Dialog/Dialog';
+import {FormApi, YTDFDialog, makeErrorFields} from '../../../components/Dialog';
 import React from 'react';
 import cn from 'bem-cn-lite';
 import PropTypes from 'prop-types';
